requirements.txt
Flask
lxml
requests
aiohttp
PyJWT
PyYAML
